Apply rule 4-3 to LAX and STRICT.
Added reference to term "RDF graph".

Updated rule 4-3 to make the union of RDF graphs for the JSON and the schema explicit.

Renamed rule 4-3 to reflect that it's not just about LAX.

Changed conformance targets on rule 4-3 to include LAX and STRICT.

Revised definition of STRICT to refer to "conformant schema document set".

Closes https://github.com/NIEM/NIEM-JSON-Spec/issues/6.